otros seeders y factories

+++ proyectodbd/database/migrations/2019_05_30_064707_create_tarjeta_credito_table.php
@@ -16,7 +16,8 @@ class CreateTarjetaCreditoTable extends Migration
         Schema::create('tarjeta_credito', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->timestamps();
-            $table->integer('numero');
+            // Cambiado de integer -> bigInteger
+            $table->bigInteger('numero');
             $table->date('fecha_expiracion');
             $table->string('nombre_titular');
             $table->string('pais_facturacion');
